Extract logic for get options from argv to a specific class
