+++ .rubocop_todo.yml
@@ -1,7 +1,12 @@
 # Offense count: 1
 Metrics/AbcSize:
   Exclude:
-    - 'lib/rubycritic/cli/options.rb'
+    - 'lib/rubycritic/cli/options/argv.rb'
+
+# Offense count: 1
+Metrics/LineLength:
+  Exclude:
+    - 'lib/rubycritic/cli/options/argv.rb'
 
 # Offense count: 2
 # Configuration parameters: CountComments.
